1.0.0-beta.8

+++ scripts/utils.js
@@ -1,9 +1,13 @@
-const fs = require('fs-extra');
-const path = require('path');
+const fs = require("fs-extra");
+const path = require("path");
 
-const dir = path.resolve(__dirname, '..', 'dist');
+const dir = path.resolve(__dirname, "..", "dist");
 
 function switchVersion(version) {
+  if (!fs.existsSync(dir)) {
+    return;
+  }
+
   fs.emptyDirSync(`${dir}/cur`);
   fs.copySync(`${dir}/v${version}`, `${dir}/cur`);
 }
